feature: skip if already exists

+++ setup.sh
@@ -7,8 +7,8 @@ function check_file_exists() {
     for file in "${DOTFILES[@]}"; do
         local home_file="$HOME/${file}"
         if [ -e "$home_file" ]; then
-            echo "Error: $home_file already exists."
-            exit 1
+            echo "Skipping: $home_file already exists."
+            continue
         fi
     done
 }
@@ -17,8 +17,12 @@ function make_symbolic_links() {
     for file in "${DOTFILES[@]}"; do
         local source_file="$(pwd)/${file}"
         local target_link="$HOME/${file}"
-        echo '🔗 Linking file' "$source_file" 'to' "$target_link"
-        ln -sf "$source_file" "$target_link"
+        if [ ! -e "$target_link" ]; then
+            echo '🔗 Linking file' "$source_file" 'to' "$target_link"
+            ln -sf "$source_file" "$target_link"
+        else
+            echo 'Skipping existing file:' "$target_link"
+        fi
     done
 }
